pyproject.toml
[tool.poetry]
name = "temple-web"
version = "0.1.0"
description = ""
authors = ["aahnik <daw@aahnik.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
django-solo = "^2.1.0"
django-json-widget = "^1.1.1"
pillow = "^9.5.0"
requests = "^2.31.0"
djlint = "^1.31.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
isort = "^5.12.0"
black = "^23.3.0"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
